mqtt-exporter/requirements.txt
paho-mqtt==1.4.0
prometheus-client==0.5.0
